Script_python/requirements.txt
pandas==2.1.3
numpy==1.26.2
PyYAML==6.0.1
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
notebook==7.0.6
scipy==1.11.4
python-dateutil==2.8.2
pytz==2023.3.post1
pytest==7.4.3
pytest-cov==4.1.0
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pandas-stubs==2.1.1.230928
types-PyYAML==6.0.12.12
pytest-mock==3.11.1

